added RegistrationServlet.java for handling c2dm registraion per user. currently just sending DevRegAndIDPairs object for testing, but need to be replaced with the proper User class.
Ran - once the class will be complete, you can just import it from the HttpRequest and add reg_id to the user by the google account.
i added the code for sending push to the RegistrationServlet.java, justs for testing. once there will be servlets that are responsible for sending diffrent types of push - this code will be copied to relevant places.